requirements.txt
fastapi[all]
fastapi-sso
pyjwt
